Changed names/added dashboard name logic
I changed the name of a few .js files so that they make more sense. CourseView is now CampaignInfo because it displays info about that campaign. TestView is now CourseInfo for the same logic.
