+++ App.js
@@ -2,10 +2,9 @@ import React from 'react';
 import {createStackNavigator} from 'react-navigation';
 import Start from'./Start';
 import CampaignView from './CampaignView';
-import CourseView from './CourseView';
+import CampaignInfo from './CampaignInfo';
 import Login from './Login';
-import TestView from './TestView';
-import PushNotifications from './PushNotifications';
+import CourseInfo from './CourseInfo';
 console.disableYellowBox = true;
 
 const RootStack = createStackNavigator(
@@ -13,11 +12,10 @@ const RootStack = createStackNavigator(
     Login: Login,
     Start: Start,
     CampaignView: CampaignView,
-    CourseView: CourseView,
-    TestView: TestView,
-    PushNotifications: PushNotifications,
+    CampaignInfo: CampaignInfo,
+    CourseInfo: CourseInfo,
   },
-  
+
   {
     initialRouteName: 'Login'
   }
