+++ users.json
@@ -1,5 +1,5 @@
 {"users":[
 	{"id":"Ryan Berry","courses":[123,234,345,456]},
-	{"id":"Lydia Trumbull","courses":[567,678]},
+	{"id":"Lydia Trumbull","courses":[567,678, 345, 123, 890]},
 	{"id": "Kyrrah Nork", "courses":[345, 456, 567, 678, 789]}
 ]}
